refactor(vouch): move vouch system out of the MCP server into GitHub CI tooling
Per scope feedback, the vouch system must not be part of the AutoMobile
MCP server — it adds no MCP tools, no DB tables, and ships nothing in the
server bundle. It is purely GitHub CI/issue gating tooling.

- Remove all DB coupling: delete the vouch SQLite migration, VouchRepository,
  VouchService, and the DB-backed repository test; revert src/db/types.ts.
  `src/` now has zero references to vouch.
- Relocate the pure domain + GitHub modules from src/features/vouch/ to
  scripts/github/vouch/ (engine, policy, types, snapshot, file store, gate,
  runner). The committed JSON graph (.github/vouch/graph.json) is the only
  store; trust state lives in git history, not a database.
- Move tests to test/github/vouch/.

Also harden the Action's REST client for the CodeQL "file data in outbound
network request" finding: validate event-derived issue numbers as positive
safe integers and pin the request origin to the fixed GitHub API host so no
event-file string can redirect the request.
